test(e2e): cover runtime summary provenance
